+++ command-center-sdk/src/navigation/index.ts
@@ -1,3 +1,4 @@
+export * from "./ApplicationImmersiveBar.js";
 export * from "./ApplicationNavigationPanel.js";
 export * from "./ApplicationNavigationShell.js";
 export * from "./ApplicationNavigationTrigger.js";